refactor(shared): rename alimtalk history/trigger types to message domain (BJJ-253)
- types/alimtalk.ts merged into types/message.ts; MessageLogRecord, MessageTrigger* family
- channel/provider setting types (AlimtalkProvider 등) kept unchanged
