+++ packages/shared/package.json
@@ -10,7 +10,6 @@
         ".": "./src/index.ts",
         "./api": "./src/api/index.ts",
         "./tokens/status-badge": "./src/tokens/status-badge.ts",
-        "./types/alimtalk": "./src/types/alimtalk.ts",
         "./types/eformsign": "./src/types/eformsign.ts",
         "./types/message": "./src/types/message.ts",
         "./types/service-record": "./src/types/service-record.ts",
