+++ packages/shared/src/index.ts
@@ -1,4 +1,1 @@
-// Intentionally empty: P5 slices 5.2-5.5 move shared code (eslint plugin,
-// lib/utils, drifted types, API plumbing) here incrementally. Each slice is
-// its own PR with the frontend/mobile test+lint baselines held.
-export {};
+export * from "./types/message";
