+++ packages/shared/src/types/message.ts
@@ -1,3 +1,5 @@
+import type { SystemTemplateKey } from "./system-template";
+
 // Shared message contracts used by both frontend and mobile.
 //
 // The source of truth for these shapes is the backend contract surface:
@@ -5,17 +7,276 @@
 // - backend/interface/dto/message-template.dto.ts
 // - backend/interface/dto/message-delivery.dto.ts
 // - backend/interface/dto/message-sender-approval.dto.ts
+// - backend/interface/dto/message-trigger.dto.ts
 // - backend/interface/dto/system-admin.dto.ts
 // - backend/domain/entities/message.entity.ts
 // - backend/domain/entities/message-template.entity.ts
+// - backend/domain/entities/message-log.entity.ts
+// - backend/domain/entities/message-trigger-rule.entity.ts
+// - backend/domain/constants/message-trigger-catalog.ts
 // - backend/interface/controllers/message-delivery.controller.ts
 // - backend/interface/controllers/system-setting.controller.ts
+// - backend/application/services/message-trigger.service.ts
 // - backend/application/services/system-admin.service.ts
 //
-// Message and message-template controllers currently return Date-backed
-// entities directly. Those dates serialize to ISO strings on the wire, so the
-// client-facing response types below use string while Raw* variants retain Date
-// for backend-reference parity.
+// Message, message-template, message-trigger, and message-log controllers
+// currently return Date-backed entities/views directly. Those dates serialize to
+// ISO strings on the wire, so the client-facing response types below use string
+// while Raw* variants retain Date for backend-reference parity.
+
+export type MessageTriggerEventType =
+  | "CLIENT_CREATED"
+  | "SERVICE_START"
+  | "SERVICE_END"
+  | "EMPLOYEE_ASSIGNED";
+
+export type MessageTriggerOffsetType =
+  | "IMMEDIATE"
+  | "SAME_DAY"
+  | "BEFORE_DAYS"
+  | "AFTER_DAYS";
+
+export type MessageTriggerRecipientType =
+  | "CLIENT"
+  | "PRIMARY_EMPLOYEE"
+  | "SECONDARY_EMPLOYEE";
+
+export type MessageTriggerTemplateKey =
+  | "CLIENT_WELCOME"
+  | "SERVICE_START_REMINDER"
+  | "SERVICE_INFO"
+  | "SERVICE_END_REMINDER"
+  | "EMPLOYEE_ASSIGNED"
+  | "SERVICE_FEEDBACK_LINK"
+  | "CLIENT_GREETING"
+  | "PRICE_INFO"
+  | "REMINDER"
+  | "THANKS"
+  | "SURVEY"
+  | "INFO";
+
+// Which system template's body each SMS trigger template renders. One source of
+// truth for both the backend SMS delivery and the form preview.
+export const SMS_TRIGGER_TO_SYSTEM_TEMPLATE: Partial<
+  Record<MessageTriggerTemplateKey, SystemTemplateKey>
+> = {
+  SERVICE_INFO: "SERVICE_INFO",
+  CLIENT_GREETING: "GREETING",
+  PRICE_INFO: "PRICE_INFO",
+  REMINDER: "REMINDER",
+  THANKS: "THANKS",
+  SURVEY: "SURVEY",
+  INFO: "INFO",
+};
+
+// Canonical source of truth for which trigger templates are delivered over SMS
+// vs alimtalk. Adding a new SMS template here flows it through the SMS form's
+// data-driven dropdowns, the channel filters, and the backend delivery drift
+// guard without hardcoded per-surface lists.
+export const SMS_TRIGGER_TEMPLATE_KEYS: MessageTriggerTemplateKey[] = [
+  "SERVICE_INFO",
+  "SERVICE_FEEDBACK_LINK",
+  "CLIENT_GREETING",
+  "PRICE_INFO",
+  "REMINDER",
+  "THANKS",
+  "SURVEY",
+  "INFO",
+];
+
+export function getTriggerTemplateChannel(
+  key: MessageTriggerTemplateKey,
+): "sms" | "alimtalk" {
+  return SMS_TRIGGER_TEMPLATE_KEYS.includes(key) ? "sms" : "alimtalk";
+}
+
+export type SupportedTriggerProvider = "aligo_alimtalk";
+export type AlimtalkProvider = SupportedTriggerProvider | "none";
+
+export interface MessageTriggerTemplateVariable {
+  key: string;
+  label: string;
+}
+
+export interface MessageTriggerTemplateProviderConfig {
+  templateKey: string;
+}
+
+export interface MessageTriggerTemplateCatalogItem {
+  key: MessageTriggerTemplateKey;
+  name: string;
+  description: string;
+  allowedEventTypes: MessageTriggerEventType[];
+  allowedRecipientTypes: MessageTriggerRecipientType[];
+  requiredVariables: MessageTriggerTemplateVariable[];
+  providers: Partial<
+    Record<SupportedTriggerProvider, MessageTriggerTemplateProviderConfig>
+  >;
+}
+
+export interface MessageTriggerRule {
+  id: string;
+  branchId: string | null;
+  name: string;
+  isActive: boolean;
+  eventType: MessageTriggerEventType;
+  offsetType: MessageTriggerOffsetType;
+  offsetDays: number;
+  recipientType: MessageTriggerRecipientType;
+  templateKey: MessageTriggerTemplateKey;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface CreateMessageTriggerRuleDto {
+  name: string;
+  isActive?: boolean;
+  eventType: MessageTriggerEventType;
+  offsetType: MessageTriggerOffsetType;
+  offsetDays?: number;
+  recipientType: MessageTriggerRecipientType;
+  templateKey: MessageTriggerTemplateKey;
+}
+
+export type UpdateMessageTriggerRuleDto =
+  Partial<CreateMessageTriggerRuleDto>;
+
+export type MessageTriggerJobStatus =
+  | "pending"
+  | "sent"
+  | "failed"
+  | "canceled";
+
+export interface UpcomingMessageTriggerJobPayload {
+  clientId?: number | null;
+  clientName?: string | null;
+  employeeId?: number | null;
+  employeeName?: string | null;
+  memberId: string;
+  recipientName: string;
+  recipientPhone: string;
+  templateVariables: Record<string, string>;
+  buttonUrl?: string | null;
+}
+
+export interface UpcomingMessageTriggerJob {
+  id: string;
+  ruleId: string;
+  ruleName: string;
+  eventType: MessageTriggerEventType | null;
+  offsetType: MessageTriggerOffsetType | null;
+  offsetDays: number;
+  recipientType: MessageTriggerRecipientType;
+  recipientPhone: string | null;
+  templateKey: MessageTriggerTemplateKey;
+  status: MessageTriggerJobStatus;
+  scheduledFor: string;
+  sentAt: string | null;
+  canceledAt: string | null;
+  cancelReason: string | null;
+  clientId: number | null;
+  employeeScheduleId: number | null;
+  payload: UpcomingMessageTriggerJobPayload;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export type MessageLogStatus = "pending" | "sent" | "failed";
+
+export interface MessageLogRecord {
+  id: number;
+  provider: string;
+  templateKey: string;
+  triggerJobId: string | null;
+  receiver: string;
+  clientId: number | null;
+  messageBody: string;
+  variables: Record<string, string>;
+  status: MessageLogStatus;
+  aligoMid: string | null;
+  errorMessage: string | null;
+  attempts: number;
+  lastAttemptAt: string | null;
+  nextRetryAt: string | null;
+  createdAt: string;
+  updatedAt: string;
+  ruleId: string | null;
+  ruleName: string | null;
+  eventType: MessageTriggerEventType | null;
+  offsetType: MessageTriggerOffsetType | null;
+  offsetDays: number;
+  scheduledFor: string | null;
+  recipientType: MessageTriggerRecipientType | null;
+  recipientName: string | null;
+  clientName: string | null;
+  employeeName: string | null;
+}
+
+export interface AlimtalkProviderResponse {
+  provider: AlimtalkProvider;
+  enabled: boolean;
+  updatedAt?: string;
+}
+
+export type AlimtalkTemplateType = "BA" | "EX" | "AD" | "MI";
+export type AlimtalkTemplateEmphasisType = "NONE" | "TEXT" | "IMAGE";
+export type AlimtalkTemplateButtonLinkType =
+  | "WL"
+  | "AL"
+  | "BK"
+  | "MD"
+  | "DS"
+  | "AC";
+
+export interface AlimtalkTemplateButton {
+  name: string;
+  linkType: AlimtalkTemplateButtonLinkType;
+  linkM?: string;
+  linkP?: string;
+  linkI?: string;
+  linkA?: string;
+}
+
+export interface CreateAlimtalkTemplateButtonDto
+  extends AlimtalkTemplateButton {}
+
+export interface CreateAlimtalkTemplateDto {
+  name: string;
+  tplType: AlimtalkTemplateType;
+  tplEmType: AlimtalkTemplateEmphasisType;
+  title?: string;
+  subtitle?: string;
+  content: string;
+  extra?: string;
+  advert?: string;
+  buttons: CreateAlimtalkTemplateButtonDto[];
+}
+
+export interface AlimtalkTemplateListItem {
+  templateCode: string;
+  name: string;
+  content: string;
+  title?: string;
+  subtitle?: string;
+  extra?: string;
+  advert?: string;
+  templateType: AlimtalkTemplateType;
+  emphasisType: AlimtalkTemplateEmphasisType;
+  inspectionStatus?: string;
+  isApproved: boolean;
+  category?: string;
+  buttons: AlimtalkTemplateButton[];
+  createdAt?: string;
+  updatedAt?: string;
+  senderKey?: string;
+}
+
+// The create-template endpoint currently returns the raw Aligo payload.
+export interface AligoTemplateCreateResponse {
+  code: number;
+  message: string;
+  info?: Record<string, unknown>;
+}
 
 export type MessageTemplateVariableType =
   | "text"
